refactor: re-position files and components
